ref(ts): Refactor `<RuleRow>` out of `projectAlertRules.jsx`
This moves `<RuleRow>` into its own module and converts it to typescript.
